Fix some PR suggestions
